4번까지 함

+++ djJava/MyHelloJava01/src/my/hello/javastudy02/Student.java
@@ -17,7 +17,7 @@ public class Student {
 	
 	//»ý¼ºÀÚ
 	public Student
-	(String name, int age, int score) {
+	(String name, int age, double score) {
 		this.name = name;
 		this.age = age;
 		this.score = score;
